+++ lib/pkgxx/environment.cxx
@@ -105,6 +105,20 @@ namespace pkgxx {
                 return vPKGSRCDIR;
             }).share();
 
+        // WRKOBJDIR
+        WRKOBJDIR = std::async(
+            std::launch::deferred,
+            [&]() {
+                fs::path vWRKOBJDIR = cgetenv("WRKOBJDIR");
+                if (vWRKOBJDIR.empty()) {
+                    if (auto v = pkgxx::extract_mkconf_var(MAKECONF.get(), "WRKOBJDIR").value(); v) {
+                        vWRKOBJDIR = v.value();
+                    }
+                }
+                _var_logger("WRKOBJDIR", vWRKOBJDIR.string());
+                return vWRKOBJDIR;
+            }).share();
+
         // WRKDIR_BASENAME
         WRKDIR_BASENAME = std::async(
             std::launch::deferred,